+++ frontend/src/modules/displays/index.ts
@@ -12,7 +12,6 @@
  */
 export * from './Carousel';
 export * from './ModeBanner';
-export * from './MultiStepWizard';
 export * from './ToggleValue';
 export * from './ValueDisplay';
 export * from './ValueSelectorDisplay';
