+++ frontend/src/modules/modals/DiscardAlarmLimitsContent.tsx
