+++ frontend/src/modules/modals/index.ts
@@ -1,0 +1,9 @@
+/**
+ * @summary ReUsable modals
+ *
+ * @file Components which can be reused across app
+ *
+ */
+export * from './MultiStepWizard';
+export * from './ModalPopup';
+export * from './DiscardAlarmLimitsContent';
